# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VRF Monitor LG - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /systems:
    get:
      operationId: listSystems
      tags: [systems]
      summary: List all VRF systems (excludes trashed)
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [escritorios_xp, espacos_xp]
      responses:
        "200":
          description: List of systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VrfSystem"
    post:
      operationId: createSystem
      tags: [systems]
      summary: Create a new VRF system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VrfSystemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VrfSystem"
...
  /systems/{systemId}:
    get:
      operationId: getSystem
      tags: [systems]
      summary: Get a VRF system
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: System found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VrfSystem"
        "404":
          description: Not found
    patch:
      operationId: updateSystem
      tags: [systems]
      summary: Update a VRF system
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VrfSystemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VrfSystem"
    delete:
      operationId: deleteSystem
      tags: [systems]
      summary: Soft-delete a VRF system (moves to trash)
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Moved to trash
...
  /systems/{systemId}/restore:
    post:
      operationId: restoreSystem
      tags: [systems]
      summary: Restore a trashed VRF system
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VrfSystem"
  /systems/{systemId}/permanent:
    delete:
      operationId: permanentlyDeleteSystem
      tags: [systems]
      summary: Permanently delete a trashed VRF system
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Permanently deleted
  /systems/{systemId}/summary:
    get:
      operationId: getSystemSummary
      tags: [systems]
      summary: Get system health summary and trend
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: System health summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSummary"
  /systems/{systemId}/technical-report:
    get:
      operationId: exportTechnicalReport
      tags: [systems]
      summary: Export an editable Word technical report
      description: >
        Generates a DOCX report from persisted system, startup, reading,
        photo and AI-analysis data. If sessionId is omitted, all sessions are
        included. No new AI analysis is performed.
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: sessionId
          in: query
          required: false
          description: Export only this reading session when provided.
          schema:
            $ref: "#/components/schemas/ExportTechnicalReportQuery"
      responses:
        "200":
          description: Editable Word technical report attachment
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid systemId or sessionId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: System not found, deleted, or session does not belong to the system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Report generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ExportTechnicalReportQuery:
      type: integer
      minimum: 1
...
    VrfSystem:
      type: object
      required: [id, code, name, category, vrfType, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [escritorios_xp, espacos_xp]
        deletedAt:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        floor:
          type: ["string", "null"]
        servedArea:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        vrfType:
          type: string
          enum: [multi_v_ii, multi_v_iii, multi_v_iv, multi_v_5]
        startupDate:
          type: ["string", "null"]
        building:
          type: ["string", "null"]
        condensationType:
          type: ["string", "null"]
          enum: [air, water, null]
        notes:
          type: ["string", "null"]
        healthStatus:
          type: ["string", "null"]
          enum: [healthy, warning, critical, null]
        lastReadingDate:
          type: ["string", "null"]
        createdAt:
          type: string
    VrfSystemInput:
      type: object
      required: [code, name, vrfType, category]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: [escritorios_xp, espacos_xp]
        location:
          type: string
        floor:
          type: string
        servedArea:
          type: string
        model:
          type: string
        vrfType:
          type: string
          enum: [multi_v_ii, multi_v_iii, multi_v_iv, multi_v_5]
        startupDate:
          type: string
        building:
          type: string
        condensationType:
          type: string
          enum: [air, water]
        notes:
          type: string
    VrfSystemUpdate:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [escritorios_xp, espacos_xp]
        location:
          type: string
        floor:
          type: string
        servedArea:
          type: string
        model:
          type: string
        vrfType:
          type: string
          enum: [multi_v_ii, multi_v_iii, multi_v_iv, multi_v_5]
        startupDate:
          type: string
        building:
          type: string
        condensationType:
          type: string
          enum: [air, water]
        notes:
          type: string
...
    SystemSummary:
      type: object
      required: [system, totalSessions, latestHealthStatus, parameterTrends, hasStartupReport, usingLgReference]
      properties:
        system:
          $ref: "#/components/schemas/VrfSystem"
        totalSessions:
          type: integer
        latestHealthStatus:
          type: ["string", "null"]
        latestAnalysis:
          type: ["string", "null"]
        parameterTrends:
          type: array
          items:
            $ref: "#/components/schemas/ParameterTrend"
        hasStartupReport:
          type: boolean
          description: True quando existe ao menos um relatorio de partida processado com sucesso para o sistema.
        usingLgReference:
          type: boolean
          description: True quando o sistema vai utilizar a tabela referencia LG como baseline padrao (condensacao a Ar e sem relatorio de partida processado).
    ParameterTrend:
      type: object
      required: [parameter, unit, values]
      properties:
        parameter:
          type: string
        unit:
          type: string
        values:
          type: array
          items:
            type: object
            required: [date, value, status]
            properties:
              date:
                type: string
              value:
                type: ["number", "null"]
              status:
                type: string
...