# lib/api-zod/src/generated/types/exportTechnicalReportQuery.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VRF Monitor LG - API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * @minimum 1
 */
export type ExportTechnicalReportQuery = number;