# lib/api-zod/src/generated/types/vrfSystem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VRF Monitor LG - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VrfSystemCategory } from "./vrfSystemCategory";
import type { VrfSystemCondensationType } from "./vrfSystemCondensationType";
import type { VrfSystemHealthStatus } from "./vrfSystemHealthStatus";
import type { VrfSystemVrfType } from "./vrfSystemVrfType";

export interface VrfSystem {
  id: number;
  code: string;
  name: string;
  category: VrfSystemCategory;
  /** @nullable */
  deletedAt?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  floor?: string | null;
  /** @nullable */
  servedArea?: string | null;
  /** @nullable */
  model?: string | null;
  vrfType: VrfSystemVrfType;
  /** @nullable */
  startupDate?: string | null;
  /** @nullable */
  building?: string | null;
  /** @nullable */
  condensationType?: VrfSystemCondensationType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  healthStatus?: VrfSystemHealthStatus;
  /** @nullable */
  lastReadingDate?: string | null;
  createdAt: string;
}