# lib/api-zod/src/generated/types/vrfSystemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VRF Monitor LG - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VrfSystemInputCategory } from "./vrfSystemInputCategory";
import type { VrfSystemInputCondensationType } from "./vrfSystemInputCondensationType";
import type { VrfSystemInputVrfType } from "./vrfSystemInputVrfType";

export interface VrfSystemInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  category: VrfSystemInputCategory;
  location?: string;
  floor?: string;
  servedArea?: string;
  model?: string;
  vrfType: VrfSystemInputVrfType;
  startupDate?: string;
  building?: string;
  condensationType?: VrfSystemInputCondensationType;
  notes?: string;
}