# lib/api-zod/src/generated/types/vrfSystemUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VRF Monitor LG - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VrfSystemUpdateCategory } from "./vrfSystemUpdateCategory";
import type { VrfSystemUpdateCondensationType } from "./vrfSystemUpdateCondensationType";
import type { VrfSystemUpdateVrfType } from "./vrfSystemUpdateVrfType";

export interface VrfSystemUpdate {
  code?: string;
  name?: string;
  category?: VrfSystemUpdateCategory;
  location?: string;
  floor?: string;
  servedArea?: string;
  model?: string;
  vrfType?: VrfSystemUpdateVrfType;
  startupDate?: string;
  building?: string;
  condensationType?: VrfSystemUpdateCondensationType;
  notes?: string;
}